Crate rusteuler

Modules

problem_1

If we list all the natural numbers below 10 that are multiples of 3 or 5, we get 3, 5, 6 and 9. The sum of these multiples is 23.

Find the sum of all the multiples of 3 or 5 below 1000.

problem_2

Each new term in the Fibonacci sequence is generated by adding the previous two terms. By starting with 1 and 2, the first 10 terms will be:

1, 2, 3, 5, 8, 13, 21, 34, 55, 89, ...

By considering the terms in the Fibonacci sequence whose values do not exceed four million, find the sum of the even-valued terms.

problem_3

The prime factors of 13195 are 5, 7, 13 and 29.

What is the largest prime factor of the number 600851475143 ?

problem_4

A palindromic number reads the same both ways. The largest palindrome made from the product of two 2-digit numbers is 9009 = 91 × 99.

Find the largest palindrome made from the product of two 3-digit numbers.

problem_5

2520 is the smallest number that can be divided by each of the numbers from 1 to 10 without any remainder.

What is the smallest positive number that is evenly divisible by all of the numbers from 1 to 20?

problem_6

The sum of the squares of the first ten natural numbers is,

1 squared plus 2 squared plus ellipsis plus 10 squared equals 385

The square of the sum of the first ten natural numbers is,

left-parenthesis 1 plus 2 plus ellipsis plus 10 right-parenthesis squared equals 55 squared equals 3025

Hence the difference between the sum of the squares of the first ten natural numbers and the square of the sum is
3025 minus 385 equals 2640
.

Find the difference between the sum of the squares of the first one hundred natural numbers and the square of the sum.

util
Utilities that can be shared across the solutions.