Module rusteuler::problem_9

A Pythagorean triplet is a set of three natural numbers, a < b < c, for which,

a2 + b2 = c2

For example, 32 + 42 = 9 + 16 = 25 = 52.

There exists exactly one Pythagorean triplet for which a + b + c = 1000. Find the product abc.

We iterate over all values of a, b, and c less than 1000. Once we find a combination where both a2 + b2 = c2 and a + b + c = 1000, we break out of all the loops with the solution.

This solution is pretty inefficient since it needs to iterate over so many numbers.

const N: u64 = 1000;

let mut solution = 0;
'outer: for a in 0..N {
    for b in a + 1..N {
        let c_squared = a.pow(2) + b.pow(2);
        for c in b + 1..N {
            if c.pow(2) == c_squared && a + b + c == N {
                solution = a * b * c;
                break 'outer;
            }
        }
    }
}

assert_eq!(solution, 31875000);

Runs in around 3.6597917840 seconds